QRawFont: de-inline two-arg advancesForGlyphIndexes()
Functions in which owning containers allocate should not be defined
inline, because of the amount of code generated.

Unlike similar changes in the past, this wasn't found by Clang
-ftime-trace, but highlighted by adding GCC -Wnrvo to headersclean.

Instead of rewriting the function, pull it behind the ABI boundary.

Can't pick back, because it adds an exported symbol (on Unix, Windows
already had it, due to the wholesale export of the class), so isn't
forwards-BC.

Change-Id: I0d7ac417e129023f9688028685e818e7ab6ab7e2
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>