QWindows11Style: Rework alternating rows ItemViews
Current itemviews which use alternating rows use the wrong text color
for selected items, also it is difficult to distinguish selected and
hovered items in itemviews with alternating rows. This patch sets the
right color for selected itemviews and is using a alpha mask to
distinguish selected and hovered items.

Fixes: QTBUG-137726
Change-Id: I1afc93e7426519a473b67b6f9b6da114218ec346
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>