QMatrix4x4: restore NRVO in op*(M,M) 2/5: de-alias m and m1
In the 2D case, we first assign m1 to m and then multiply m2 into it.

In the first three steps, we're using the fact that, at this point,
and for the fields referenced, m is still the same as m1.

This causes a dependency ordering of these three lines with respect to
the following three lines, though, making it both harder to read for
humans and harder to optimize for compilers. It also prevents us from
writing to m in memory order.

So don't read from m on the RHS, read from m1.

Amends 13b3545e833f6175f686c9776e1510db3f3f11eb.

Pick-to: 6.11 6.10 6.8
Change-Id: I8790974cb3337470d9359d09a943527a75f99527
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>