QMatrix4x4: restore NRVO in op*(M,M) 4/5: use Qt::Uninitialized ctor
Instead of copying the LHS operand into the result and selectively
overwriting fields whose values are not statically known for scaling
and translation, write the known values into those fields manually.

This allows us to use the Qt::Uninitialized constructor, like in the
3D case, which, in the last step, will allow us to re-enable NRVO in
this function.

It should also improve numerical stability, because it resets what
might have become merely fuzzily-0/1 values to known, clean ones.

Amends the start of the public history.

Pick-to: 6.11 6.10 6.8
Change-Id: I595498787cebe069c36c1797a9dbe1a94780066d
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>