macOS: Use non-native file dialog if sandboxed app is missing entitlement
If a sandboxed app is missing the user-selected read-only or read-write
entitlement the native file dialogs will fail to open (and will break
in the debugger via REPORT_APP_ENTITLEMENTS_INSUFFICIENT if running
in Xcode).

The non-native file dialogs still work, but will just show the files
we have access to through the app's bundle and sandbox. But it's better
than nothing.

Fixes: QTBUG-141415
Pick-to: 6.10 6.8 6.5
Change-Id: Ia2007e5c03df73156bc4e2a2badf8a7145a83ded
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>