Improve QPainter::drawTiledPixmap() docs
This provides a more explicit description of the mental model behind the
QPainter::drawTiledPixmap() API, and also sheds some light on why the
position argument is in the device independent pixels (because it's
inside the specified rectangle), and not the device pixels.

Behind the scenes, the position is still wrapped around the pixmap, but
it's rather an implementation detail.

Change-Id: I709e66a7abc535652a3f57134bc34aa41c3ba013
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>