a11y: Set buddy in print dialog's "Copies" panel
Set the combobox labelled by the "Page Set" label as
its buddy, which also makes sure that the accessible relations
between the label and the combobox are set.
This makes screen readers like Orca on Linux announce the
label text when the combobox receives focus, so the
meaning of the combobox becomes clear.

Change-Id: I7f82b07e741b743268c4b6817ceed671ee2657c6
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>