a11y: Set buddies in print properties dialog's "Page" tab
Set the comboboxes labelled by the "Pages per sheet"
and "Page order" labels as their buddies, which also
makes sure that the accessible relations
between the label and the combobox are set.
This makes screen readers like Orca on Linux announce the
label text when the combobox receives focus, so the
meaning of the combobox becomes clear.

Change-Id: Ibf0df6970d7f80f047751be29cbe01ab796a377b
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>