a11y: Set buddies in print properties dialog's "Job Options" tab
Set the widgets labelled by labels in the "Job Options"
tab of the printer properties dialog as their buddies, which
also makes sure that the accessible relations between the
label and the widgets are set.
This makes screen readers like Orca on Linux announce the
label text when the widget receives focus, so the
meaning of the widget becomes clear.

Change-Id: I096d1e835b8d3aa9a081436c256933becdb6872a
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>