a11y: Align visual and focus order in print dialog's "Job Options" tab
By default, the tab focus order matches the order in which
widgets are created.

While that could manually be overriden, just define the
widgets in the .ui file in their visual order, to ensure
that the "Start" combobox receives focus before the
"End" one below it when using the Tab key to navigate
through the dialog.

Change-Id: I362bd90df22dc7b3b7ebab1e7c704952bd3a7169
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>