a11y: Set buddies in print properties dialog's "Advanced" tab
Set the comboboxes for the advanced PPD options
in the "Advanced" tab of the printer properties dialog
as the labels' buddies, which also makes sure that the
accessible relations between the label and the widgets are set.
This makes screen readers like Orca on Linux announce the
label text when the combobox receives focus, so the
meaning/context becomes clear.

Change-Id: I1b2b9f4ba682577673d97de66f5ad2fb0cc6ce83
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>