a11y: Align visual and focus order in print dialog's "Page" tab
By default, the tab focus order matches the order in which
widgets are created.

While that could manually be overriden, just define the
widgets in the .ui file in their visual order, to ensure
that the combobox to set the unit to use receives focus
before the "Page Size" combobox below it when using the
Tab key to navigate through the dialog.

Change-Id: Ib6130b71c80048019e7874e695a7767225da7cfa
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>