tst_Moc: port relatedMetaObjects{WithinNamespaces,NameConflicts}() away from Q_ENUMS
Like relatedMetaObjectsInGadget() in a previous commit, these tests
doesn't specifically check Q_ENUMS, but e.g. the use of another
QObject's enum in a class not derived from the the first QObject.

So port from Q_ENUMS to Q_ENUM, in order to help with the Q_ENUMS
deprecation effort.

Task-number: QTBUG-99060
Pick-to: 6.10 6.8 6.5
Change-Id: Ibdb58df188ab2fcbb15f3d1d13998e271180b564
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>