A11y Windows: report nested top-level window/widget
Amends 9b01520d05beafe39d199298c155217e587a9312.

Previously, when a nested window or dialog was shown without a focused
widget, it was not reported to NVDA.

With this update, if the widget is a top-level window and has no focused
child widget, it will now be reported to NVDA as long as it is in the
active state.

Task-number: QTBUG-90897
Task-number: QTBUG-90899
Pick-to: 6.10 6.8
Change-Id: I5bb1cde178f76b65bffaba7293daa75a7a66506a
Reviewed-by: Ingo Klöcker <dev@ingo-kloecker.de>
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>