wasm: Refactor Promise
Switch from a function based implementation to making
Promise a object.

Maps straightforward to a javascript promise, except for
this:

We need to do cleanup when the promise has been fully processed.
We do this by adding a final handler.

However, the handlers are called in the order they are added, so
the final handler we add in the constructor will be executed
too early.

We solve this, in a non optimal way, by adding a final handler
after each then, and catch, handler.

This makes the interface easy to use, even if it is suboptimal

Fixes: QTBUG-142138
Change-Id: I6be3f102e838467cc98f4a9825fd5f6fe0e1e1a2
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>