QLocale/ICU: fix -unity-build
GCC says that CaseConversion::Upper is ambiguous between
- struct QUnicodeTables::CaseConversion
- enum class QtIcuPrivate::CaseConversion

In a -unity-build, those two entities are in-scope at the same time,
so disambiguate by renaming the entity with narrower scope.

The alternative, fully qualifying the enum uses with the QtIcuPrivate
namespace would have caused similar churn, and would have run the risk
of being removed again by the mythical maintenance programmer, due to
use of "using namespace QtIcuPrivate;" in at least some of the
functions that refer to the enum.

Amends eaf77edebdf1c0b1682c28c014f2d81479af0051, which added the
conflicting QUnicodeTables::CaseConversion and/or
6f4d64b21ec90a792c7c8308a6615ca87b9ea6eb, which added
QtIcuPrivate::CaseConversion, which this patch changes now.

Pick-to: 6.10 6.8 6.5
Task-number: QTBUG-109394
Change-Id: I1f6379a200ae0cea5befd5d6cf210be14c0344f0
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>