QColorSpace: add a missing std::move()
Coverity complained that the assignment of ownedIccProfile to
d->iccProfile was a copy when it could have been a move.

It's correct, so add the move(), after marking the to-be-moved-from
variable as non-const.

As a drive-by, replace unwarranted direct-initialization (the
QByteArray copy constructor isn't explicit) with more readable
copy-initialization.

Amends fc5208abc1deef1b6e206a2f6452bc6240dac195.

Pick-to: 6.10 6.8
Task-number: QTBUG-141942
Coverity-Id: 897407
Change-Id: I43e0baa2e37e1140ba51fede48bc1ee3f4b29537
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>