Add check for -Wcharacter-conversion support
Use __has_warning() to check if -Wcharacter-conversion
is supported before disabling it, in order to avoid
"unknown warning" warnings.

This is/was a problem when building with Emscripten 4.0.7,
which uses a git snapshot of Clang 21 which does not
support the warning.

Amends f20f12baab1f40ec0aad2635186e7d270139509b.

Pick-to: 6.10
Change-Id: I4c08e69013b36b87709295d9a3a79d8986078ef1
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>