q23:expected: namespace TL_ macros with Q23_
Guard our copy of tl::expected against use of a different version of
the class in projects that use both. On the C++ side, we already did,
by renaming the namespace from tl to q23. But macros are not
namespaced, and the implementation doesn't clean up its many macros at
the end of the file, so the alternative, simply undef'ing them at the
end, doesn't work, because we can't rely on "the other version" to do
the same.

So grab the nettle and prefix _all_ TL_ macros with Q23_, incl., and
in particular, the header guard, which might have caused FTBFS if a
project was happily using tl::expected and Qt now injects
q23::expected instead of tl::expected into the same header guard.

Leave the 0005 patch slot free for an eventual amendedment of
4e9f4d5d02e29b4522540b8f9b9b01b7e57a6b54.

[ChangeLog][Third-Party Code] Fixed potential compile errors in
projects that use both the original tl::expected and Qt private
headers.

Amends 6f319847d01f6fa511397b83e267d6f7a56ff322.

Fixes: QTBUG-142232
Pick-to: 6.11 6.10
Change-Id: Ib5762ec8ebe81e0c750da84be29531b7179c5025
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>