tst_qxp_is_virtual_base_of: fix a test towards ambiguous bases
If a derived class inherits from the same base class both virtually and
non-virtually, our implementation of the is_virtual_base_of trait cannot
properly detect this due to the ambiguity, and will answer "false". On
the other hand the version shipping in C++26 will correctly answer
"true" (there *is* a virtual inheritance path).

Amend the test to take this into account.

Practically speaking, this behavioral difference isn't a cause of
concern: this trait is used to guard conversions of pointers from
derived classes towards base classes. That implies that the inheritance
is not ambiguous (otherwise the pointer wouldn't be convertible to begin
with).

Amends 35878fa924d0598b1cdec5ba358e8488f6774676.

Task-number: QTBUG-138246
Pick-to: 6.10 6.8
Change-Id: I5b115a826202d1fb6c793678d4660c847c1a7c71
Reviewed-by: Marc Mutz <marc.mutz@qt.io>