QFont: fix operator<() pt.2: comparing QMaps
While the first patch fixed the obvious issue of lexicographically
comparing an unordered container, this one might or might not fix
something.

But this author requires the peace of mind of having a known-good STL
algorithm do the heavy lifting instead of being forced to analyze a
raw loop, so this patch replaces said raw loops (and the suspicious
size check that preceded them) with std::lexicographical_compare()
calls, which, despite its name, implements op< semantics, and is
proven by way-more-capable programmers than yours truly (i.e. Alex
Stepanov and David Musser) to yield a strict weak ordering when the
parametrized comparator is one (which it is, in this case,
recursively, as std::pair's op< is also defined to use lexicographical
ordering).

As such, it's only now that I feel comfortable in saying that this
operator is a strict weak ordering again.

[ChangeLog][QtGui][QFont] Fixed a Qt 6.7 regression in the stability
of the less-than operator. If you use QFont as keys in a
QMap/std::map, or otherwise use QFont ordering (equality is ok), then
we strongly recommend to update.

Amends 6160ea45b689e9d26795a18f155053ac4dc4dd6b.

Pick-to: 6.11 6.10 6.8
Fixes: QTBUG-142246
Change-Id: I1c96289af9326c7020f67674c148beca78791304
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>