QFlatMap: restore NRVO in do_take()
The usual problem, the usual solution...

At the time of construction of the return object it is known which of
the two variables to construct, so NRVO would be permissible, but most
compilers don't enable it because the structure isn't

    T result; // first (non-trivial) variable declared
              // must be a function's top-level scope
    ~~~
    return result; // only one
                   // or else all return statements are token-by-token the same

The usual fix is to wrap the tail part of the function (= the one that
constructs a T and returns it, as opposed to the part that returns
temporaries) in an IILE expression: the lambda then _has_ said
structure, so is NRVO'ed and the call to the lambda is RVO'ed in the
caller.

Invert the polarity of the if statement in order to keep the meat of
the function's git history intact.

Amends deddafe0a6a32aa438cc36c7dcfae8c323274487.

Pick-to: 6.10 6.8
Change-Id: I3f22665daca320be283e4088cf1062f115cd49e4
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>