JNI: document the JObject type, remove TECH_PREVIEW_API tags
The JObject(Base) types have been around for a while now, and are
heavily used in Qt code. We can be confident that they work. Add
documentation for the JObject helper type.

The JObjectBase type has no public usage, it's just a technicality that
takes care of type-independent implementations. Document it as internal.

Fixes: QTBUG-141156
Change-Id: Ia6567e10dc57a022beb3d3aff0420b2b9b9f6b2f
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>