QIORing/linux: split vectored read
Because Linux internally has a max read/write size for all sorts of
reads and writes (even vectored), we have to specially handle the case
where the user is reading or writing more. A previous patch already
handles the single-span read/write, this patch deals with the vectored
version.

Because a single span of bytes in the 'vector' may be large enough by
itself to surpass this internal limit we sometimes have issue, behind
the scenes, a regular write for parts of the span, but otherwise try
to write as much possible with vectored operations.

The vectored read and write part share a function that deduce how each
operation can do, and at what offset it must start etc.

Pick-to: 6.11
Task-number: QTBUG-136763
Change-Id: If6cd1f95a3cc3c9f5a38feb898c1fa686efc72a9
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>