Silence missing deployment qtpaths warning on non-Windows platforms
In 36cee508a088e1570e1d89e060c3ccd14585d3f1 we started showing a
warning when a qtpaths executable or shell wrapper was not found in
the qt prefix.

This warning would appear on all platforms equally, even though in the
end the __QT_DEPLOY_TARGET_QT_PATHS_PATH variable is only used when
deploying a Windows app using windeployqt.
An example of when the warning appears is with a yocto SDK, where the
target qt prefix does not contain a qtpaths executable.

Make sure the warning is only shown when targeting the Windows
platform.

I considered searching for candidate qtpaths also in the host prefix
when available, but such a qtpaths most likely will not show
information for the target platform, and thus deploy the wrong
libraries.

Amends 36cee508a088e1570e1d89e060c3ccd14585d3f1

Pick-to: 6.8 6.10
Fixes: QTBUG-142336
Change-Id: If3df1f61cf1550b7271268c71a5826ed673a62fa
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>