Short live QtPrivate::fuzzyCompare()!
Extract Method QtPrivate::fuzzyCompare() from the qFuzzyCompare()
implementations that correctly validate the preconditions of the
public qFuzzyCompare() (QMarginsF, QSizeF, QPointF), so we can more
easily apply the same technique to other in-tree callers.

In the process, replace comparison to literal 0.0 with calls to
qIsNull() so we're a) insulated from -Wfloat-compare and b) have a
customization point to extend to types other than float and double.

Amends:
- 473d06970d224b202e7a8ee8feaa2a2d98d5b257 (QMarginsF)
- fa0d77e290f5ccb5afa7d02716f8726aa6b810e6 (QPointF)
- 3ca9877f8cae41e1b1cb3c72d4a7872d97fb8254 (QSizeF)

Pick-to: 6.10 6.8 6.5
Task-number: QTBUG-142020
Change-Id: Ib7ec06822f8006771a1c3a96145e98d574a29fbe
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>