QQuaternion: fix remaining qFuzzyCompare() precondition violations
qFuzzyCompare() requires that neither argument is numerically
zero. This cannot be guaranteed for the quaternion elements.

Fix by using the new QtPrivate::fuzzyCompare() function, which does
things in the correct way.

As a drive-by, put the operators at the beginning of continued lines,
as requested by https://wiki.qt.io/Qt_Coding_Style#Line_breaks Item 2.

Amends the start of the public history.

Pick-to: 6.10 6.8 6.5
Change-Id: Ia317aa260ffc9fb04c10c24517a1ca3be58a1b05
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>