QPainterPath: fix remaining qFuzzyCompare() precondition violations
qFuzzyCompare() requires that neither argument is numerically
zero. This cannot be guaranteed for the positions of QPP::Elements.

Fix by using the new QtPrivate::fuzzyCompare() function, which does
things in the correct way.

I note that many of these could just delegate to the QPointF
qFuzzyCompare() overload, which, however is not available in 6.5, so
that cleanup has to come as a follow-up.

As a drive-by, in the Element equality operator, fix formatting and
pull the type comparison first, as it's cheaper (so we'll fail
faster).

Amends the start of the public history.

Pick-to: 6.11 6.10 6.8 6.5
Change-Id: I9548740a1291f56fadf39e7a6471602045d1bf4d
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>