QQuaternion: fix qFuzzyCompare() precondition violation
qFuzzyCompare() requires that neither argument is numerically
zero. One user of this function cannot guarantee this, because, while
they check for zero, they do so _after_ calling qFuzzyCompare() with
that value, so the check comes too late.

Fix by swapping the two checks around (so we check for == 0 before we
check for == 1).

Amends 8c1532eeaae6a220569cef3476d1124399478285, which changed from
qFuzzyIsNull(x - 1) (which doesn't have the problem) to
qFuzzyCompare(x, 1) (which does).

Pick-to: 6.10 6.8 6.5
Task-number: QTBUG-142020
Change-Id: I054449b5510217f83479c5a028986c569f54084c
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>