Fix failing tst_QComboBox::popupPositionAfterStyleChange openSuSE 16
tst_QComboBox::popupPositionAfterStyleChange was failing consistently
where the mouseEvent was sent just a few pixels out of the popup. Set
a fix size for the popup.

On openSuSE, the popup width appears to be twice its actual size. The
cause of this is still unclear.

Workaround to Fix consistently failing
tst_QComboBox::popupPositionAfterStyleChange on openSuSE 16.

Fixes: QTBUG-141772
Change-Id: I9f659c02ea3af862b57025aead8cc27462bf59aa
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>