WindowsStyle: Fix handling for SH_UnderlineShortcut
Underlining the shortcut key when pressing the 'ALT' key did not work
correctly - the widget did not get a repaint request and therefore the
underline was not painted.
The test if a widget needs a repaint checked SH_UnderlineShortcut but
it always returned true due to the fact that 'seenAlt' already contained
the widget itself.

Fixes: QTBUG-133528
Pick-to: 6.11 6.10 6.8
Change-Id: I35b889c8e332aa9836e0938af956bf63d3fbffd8
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>