QRMA: hook inserted objects up to auto-connect
Auto-connecting properties is enabled in models where all item data is
backed by the same QObject subclass. When assigning a new row, an entire
branch of rows in a tree, or assigning a fresh range as the children of
a row, then the new items need to be connected.

Some of these items might be nullptr, in which case we have so far
stopped the connection loop early (by returning false from the helpers).
Fix that to only stop early if a connection failed (i.e. if role names
and properties in the objects are not aligned), but continue if we
encounter a nullptr entry in the item data.

Change-Id: I2c4b5e5beedc7b38c40ee459c2e0437568b9b087
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>