QRM: implement itemData in terms of multiData
No need to duplicate the entire logic. The only overhead is that we have
to create a QModelRoleDataSpan for all roles, except our internal roles.
Using a QVarLengthArray makes that cheap.

The only special case we keep is for items that are backed by a type
that is directly convertible to a QMap<int, QVariant>. This comes with
a small overhead for acessing the element at runtime.

Pick-to: 6.10
Change-Id: I33f86d231bef9036a7e83584d1bf013f129b8075
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>