wasm: draganddrop Make sure dropaction is correct in all cases
We do not necessarily run the functions in QWasmDrag, but might
instead run the QSimpleDrag object.

So we cannot rely on m_dragState->dropAction, but have to get
the dropAction from QSimpleDrag.

To make QWasmDrag work we call setExecutedDropAction in
the cases where it is needed.

Also, do not retry the drag operation if jspi && IgnoreAction.

Fixes: QTBUG-142512
Change-Id: I2f1548db50a064033d1071a07cc42f1488610905
Reviewed-by: Lorn Potter <lorn.potter@qt.io>