QRMA: clean up memory in test case
The model doesn't take ownership of the item objects, so we have to
delete them when the row items get destroyed. As usual, tree items that
are not move-only are a mess, so make them move only, and implement
move-semantics to explicitly clear the moved-from row's array (as moving
from a std::array doesn't clear the source).

Adapt the test case accordingly.

Pick-to: 6.11
Change-Id: I4722be0bbe3846f7e274583d19b75beb502b7095
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>