QHttp2Stream: add Configuration with downloadBuffer option
Especially useful for clients who don't use the PUSH_PROMISE
functionality and don't require buffering the data.

We add the QHttp2Stream::Configuration struct to allow customizing
stream behavior. Every stream comes with a configuration and the default
constructed objects represents that. The configuration is per-stream
and does not change during the lifetime.

Add a testcase with further extensibility in mind for further stream
configurations.

Task-number: QTBUG-142473
Pick-to: 6.11 6.10
Change-Id: I1e862f4996baa61f024f40516f74fc052a9b57c4
Reviewed-by: Alexey Edelev <semlanik@gmail.com>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>