Revive QWidget on eglfs by testing for OpenGLOnRasterSurface
Removing the RasterGLSurface type seems to have a side effect
for eglfs, although the root cause is pre-existing, strictly
speaking.

Widget applications stop working completely due to makeCurrent()
and swapBuffers() complaining about the type being RasterSurface.

Previously this was RasterGLSurface and so there were no complaints.

Now that the type for top-level widget windows is RasterSurface, and
considering that a platform like eglfs is still going to render
with OpenGL onto such a surface regardless what surfaceType says,
the test QOpenGLContext uses for those warnings and bail outs should
be expanded.

Conveniently enough, there is an existing OpenGLOnRasterSurface
capability in QPlatformIntegration, reporting exactly what we
want, eglfs being the platform reporting it as true. Use it.
Technically this should have been there before, but RasterGLSurface
masked this inconsistency until now.

Change-Id: I2523cd79e480571a26af9d31a1a431f71db078d6
Fixes: QTBUG-142538
Pick-to: 6.11
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>