QThread/Unix: remove setting of the service level with itself
Calling QThreadPrivate::setQualityOfServiceLevel() with d->serviceLevel
just sets it with itself. That function does nothing on OSes other than
Apple's. On Apple's, as the comment said, we have already set the
service level in the pthread_attr_t before starting the thread.

I guess this was done for future-compatibility, but there isn't anything
on the horizon to need this. So simplify the code. Amends commit
6013f82c94c12596670cf03efe451cdd3d5713be.

Pick-to: 6.11 6.10
Change-Id: Id409a95168f17457a432fffd26b76d364f997e4e
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>