QPainterPath: fix qFuzzyCompare() use in trimmed()
qFuzzyCompare() requires that neither argument is numerically
zero, but none of the four qFuzzyCompare() calls in trimmed() can
guarantee this.

For the first two, while the function checks for f1 - f2 == 0, it does
so _after_ calling qFuzzyCompare() with that value, so the check comes
too late. It also formulates the check as qFuzzyCompare(f1, f2), which,
if f1 or f2 are zero, may fail for the same reason.

Fix by moving the equality check to be the first of the two, using
qFuzzyIsNull(f1 - f2) to avoid the problem of f1 == 0 or f2 == 0.

The remaining two calls also need to deal with f1 or f2 being zero, so
just use the new QtPrivate::fuzzyCompare() function, which handles zero
correctly.

Amends 91f502a7d65d76f4a376903c4058385a62dad277.

Pick-to: 6.11 6.10
Task-number: QTBUG-142020
Change-Id: I30940bb62a6f27d1edb07d05e726c3fe5ae12d87
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>