QRM: respect ADL in begin/end/size/swap calls
Our QRangeModelDetails::begin/end helpers explicitly tested for the
return type of std::begin/end, which disabled ADL. Similarly, we used
fully qualified std::size and std::swap to test whether those functions
are available.

The standard documents that all those functions may be overloaded,
which allows customizing the behavior of e.g. std::ranges::size etc. We
want to respect such customizations.

We can pull std::begin/end/sizeinto the Details namespace, and use
unqualified versions of those functions to enable ADL in the respective
tests. And since qSwap does that already for us, use that directly.

Rename QRangeModelDetails::begin/end to adl_begin/adl_end so that we
can use the unqualified begin/end from those, and adjust all call sites.

Add a simple test to confirm that for a custom type, custom overloads
of begin and size are called. We don't use end in a flat model, and
std::rotate might not use swap, so we can't test for those.

Pick-to: 6.11 6.10
Change-Id: Ic6df5dcd3e2e9b6b83fe100d1fdaf4bc15e70ba9
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>