QRMA: try to constrain the data source for insertColumn(s)
insertColumn can insert a data element, or a range of data elements,
which then inserts one element from the source range for each row.

insertColumns can insert a range of data elements (which is then the
data inserted into each row), or a range of ranges of data elements (in
which case we insert one range from the source for each row).

Try to constrain this by inspecting the input data type, and allowing
only compatible data elements, as well as ranges holding compatible
data elements.

Make the API test more specific - test with lists of integers, so that
we can confirm that such input data wouldn't be allowed to add columns
to a model holding pointers.

Pick-to: 6.11
Change-Id: I1b67510b3f70f147530a3ae453b01328f2e639d2
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>