CMake: Silence cmake generator warning for iOS test projects
A developer might want to open a qt auto test project to build and
launch it on an iOS simulator or device. For iOS apps we require the
Xcode generator to handle the code signing and launching.

But Qt warns when using a generator other than Ninja for a project
that loads the BuildInternals package.

Add an exception to silence the warning when building for iOS
in standalone test projects which also load the BuildInternals
package.

Pick-to: 6.8 6.10 6.11
Change-Id: I4748060238678a103578c76ce3994dc1abc7a779
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>