QEasingCurve: fix qFuzzyCompare use in compareEquals()
qFuzzyCompare() requires that neither argument is numerically zero.
Since the setters of the values that are being passed here don't
reject zero, the equality operation apparently needs to work on zero
values.

Fix by using the new QtPrivate::fuzzyCompare() function, which does
things in the correct way.

This marks the last vulnerable qFuzzyCompare() use in QtCore. I'm
closing QTBUG-142020 and will spawn off per-module tasks for the rest.

Amends the start of the public history.

Pick-to: 6.11 6.10 6.8 6.5
Fixes: QTBUG-142020
Change-Id: Ie5727b1acc419853365422efb02e311a8b89bfb4
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>