CMake: Adjust PCH generation for linked private Qt modules
QtMultimediaQuick links against QtMultimediaPrivate. Since
QtMultimediaPrivate is an interface library, we did not add its master
header to the precompiled headers of QtMultimediaQuick. However, it's
likely that if one uses a private module then headers of the public
module are used.

Now, we don't discard all interface libraries in
qt_update_precompiled_header_with_library but
- detect if it's a private Qt module
- retrieve the public Qt module target
- call qt_update_precompiled_header_with_library with that

Fixes: QTBUG-142078
Change-Id: Ic0eff21b8265cd622202efdd4ffd878e1f1fdea1
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>