rhi: vulkan: avoid clear values when no attachment clears
When a render pass targets a PreserveColorContents or
PreserveDepthStencilContents render target, there may not be any
attachment to clear. In this case, Vulkan expects a
rpBeginInfo.clearValueCount of 0, instead of the number of attachments.
Otherwise, validation layers emit the following warning:
  vkCmdBeginRenderPass(): This render pass does not have
  VkRenderPassCreateInfo.pAttachments->loadOp ==
  VK_ATTACHMENT_LOAD_OP_CLEAR but VkRenderPassBeginInfo.clearValueCount
  > 0. VkRenderPassBeginInfo.pClearValues will be ignored and no
  attachments will be cleared.

To fix this, avoid specifying clear values when none of the render pass
attachments use VK_ATTACHMENT_LOAD_OP_CLEAR (or stencilLoadOp == CLEAR).

Drive-by, remove unnecessary float casts, and add a Q_ASSERT verifying
the clear value count matches the number of attachments.

Pick-to: 6.11 6.10
Change-Id: Ib876c6b8dd943a62e1074339c5cda9b4cf414336
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>