configure: Fix "configure -developer-build -nomake doc-snippets"
Configuring a developer build with -nomake doc-snippets was still
building the doc snippets, because the feature value was set to ON.

QT_BUILD_DOC_SNIPPETS is initialized with FEATURE_doc_snippets and
FEATURE_developer_build. We use this variable for the doc_snippets
feature's AUTODETECT value. This way, we have consistent values for
QT_FEATURE_doc_snippets and QT_BUILD_DOC_SNIPPETS.

Pick-to: 6.11
Change-Id: I108b02ca2c548706d0fdd538f6ae6b7d1704fadc
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>