QRegularExpression: add missing const to PcreJitStackFree::op()
Function object's function-call operators ought to be marked as const,
so do it.

Found while working on reviewing the code with an eye to fix
QTBUG-142551.

Not picking all the way since this clearly doesn't fix a bug (the
object is stowed away inside of a std::unique_ptr, which obviously
doesn't care about the missing const, or this wouldn't have compiled
before).

Pick-to: 6.11 6.10 6.8
Change-Id: I62649504d73192519a08684701d08483dfc09f6c
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: David Faure <david.faure@kdab.com>