QtTestLib: activate initmain test case
That test was never added to the build system, and following the
simplification of the initMain detection it would no longer build. Add a
CMakeLists file and add that to the testlib auto tests, and update the
compile time test to match the new logic.

Amends 58faca4e07011510388df94ddea2f95b56750bae.

Pick-to: 6.11
Change-Id: I98b46351c8d681fd628b4793e07a5ef69bbb6b31
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>