QTabBar: fix behavior and tests when at least one scroll button is left
Currently all styles draw the scroll buttons for a QTabBar on the right
side of the tabs. The windows11 style draws them on both sides. This
revealed some small issues within the tabbar code.

QTabBarPrivate::scrollTabs() was using the wrong offset to check if the
tab is currently fully visible. The scrollRect was already adjusted some
lines above so it fits to the 'tabs' coordinate system.

During a pixel-exact wheel event (triggerd through the autotests), the
minimum scroll offset could be below zero when there are buttons on the
left which was not honored here.

Pick-to: 6.11 6.10 6.8
Change-Id: I32c9731d8b896403922324ac2e109cd7d441d850
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>