QDirListing: prevent unintended QList detach
The QStringList being iterated over is almost certainly still shared
with the caller, because this is part of the constructor call, when
the caller's argument is still alive, even if a temporary. So the
ranged for loop is force-deep-copying it just to iterate over it.

Fix by adding std::as_const(), as usual.

Amends c39a0d1e8956e042139ce3065681e4c5d07412f3.

Pick-to: 6.11 6.10 6.8
Change-Id: I31df5e3d388ef8133a667f7c5b474ae5126c2260
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>