QStyleSheetStyle/QAbstractButton: fix iconSize property for buttons
The iconSize property in QStyleSheetStyle can depend on the check state
of the button, therefore QStyleOption::state needs to be filled
accordingly. Since there is no virtual initStyleOption() in
QAbstractButton we have to add virtual
QAbstractButton::styleButtonState() to retrieve the correct button state
from the derived classes. With Qt7 this should be fixed with a virtual
initStyleOption() function.

Pick-to: 6.11
Fixes: QTBUG-142121
Change-Id: I3d035ffb403ae1f03a2e1a2ed9b016fc861ee7be
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>