QFileSystemModel: don't copy the QFileSystemModelSorter around
Found by Coverity, which complained that 'ms' should have been moved
into std::sort() instead of copying.

But the copying wouldn't stop there. Predicates are supposed to be
cheap to copy and stateless, so they can be copied around inside
algorithms as often as needed. QFileSystemModelSorter contains a
QCollator, an implicitly-shared type, so it is anything but easy to
copy.

Instead of rewriting QFileSystemModelSorter to hold a QCollator by
reference, just use std::cref() to pass it around by reference
instead. Mark it as const to statically assert that it has no mutating
state, either, and add a scope around the std::sort() to indicate that
the predicate isn't reused later.

Amends 197da3d220fb267d127288e109e691d832ce290b, which added the
QCollator to the predicate which before _was_ easy to copy (containing
just an int).

Coverity-Id: 444176
Pick-to: 6.11 6.10 6.8
Change-Id: I1c97d646f87995dc6b5ced96e724993937e9395d
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>